Alignment of Franca and DBus files, test done for POI APIs

+++ src/poi-service/poi-common/poi-common-data-model.h
@@ -41,7 +41,7 @@ typedef struct
 {
     double latitude;
     double longitude;
-    int altitude;
+    double altitude;
 } geocoordinate3D_t;
 
 enum ATTRIBUTE_LIST{
